backend/requirements_essential.txt
# Smart BI Platform - Essential Requirements
# Core packages needed to run the FastAPI backend

# FastAPI Core
fastapi>=0.104.0
uvicorn>=0.24.0

# Configuration & Environment
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Database (PostgreSQL)
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# Authentication & Security
python-jose>=3.3.0          # JWT token handling
passlib>=1.7.4              # Password hashing
python-multipart>=0.0.6     # Form data parsing

# Data Processing (Required for full application)
pandas>=2.0.0
numpy>=1.24.0

# File Uploads
python-multipart>=0.0.6

# Optional: Install separately as needed
# AI/ML: langchain groq openai
# Advanced Data: pandas numpy matplotlib plotly
# Background Tasks: celery redis
# Caching: redis
# Testing: pytest httpx